Numeric types downgrading to cope with out of memory issue

+++ ExtractIndirectCoupling/ProjectParser/BackwardMetrics.cs
@@ -4,23 +4,23 @@ namespace ProjectParser
 {
     public class BackwardMetrics : ICloneable
     {
-        double bavg = 0;
-        double bmax = 0;
-        double bmin = double.MaxValue;
-        double bcnt = 0;
-        double bsum = 0;
-        double bcntproc = 0;
+        int bavg = 0;
+        int bmax = 0;
+        int bmin = int.MaxValue;
+        int bcnt = 0;
+        int bsum = 0;
+        int bcntproc = 0;
 
         public BackwardMetrics()
         {
         }
         
-        public double Bavg { get => bavg; set => bavg = value; }
-        public double Bmax { get => bmax; set => bmax = value; }
-        public double Bmin { get => bmin; set => bmin = value; }
-        public double Bcnt { get => bcnt; set => bcnt = value; }
-        public double Bsum { get => bsum; set => bsum = value; }
-        public double Bcntproc { get => bcntproc; set => bcntproc = value; }
+        public int Bavg { get => bavg; set => bavg = value; }
+        public int Bmax { get => bmax; set => bmax = value; }
+        public int Bmin { get => bmin; set => bmin = value; }
+        public int Bcnt { get => bcnt; set => bcnt = value; }
+        public int Bsum { get => bsum; set => bsum = value; }
+        public int Bcntproc { get => bcntproc; set => bcntproc = value; }
 
         public object Clone()
         {
